+++ src/simmate/apps/quantum_espresso/workflows/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from .static_energy import StaticEnergy__QuantumEspresso__Quality00